Move param tree SubjectItem to params package
